all: replace ulong typedef with uint32_t
Replace all instances of ushort within the 3DMMEx source with uint32_t. This
should also help building on platforms with different type sizes from x86
Windows.

+++ bren/bwld.cpp
@@ -747,7 +747,7 @@ bool BWLD::FClickedActor(long xp, long yp, BACT **ppbact)
 /***************************************************************************
     Assert the validity of the BWLD.
 ***************************************************************************/
-void BWLD::AssertValid(ulong grf)
+void BWLD::AssertValid(uint32_t grf)
 {
     BWLD_PAR::AssertValid(fobjAllocated);
     AssertPo(_pgptWorking, 0);
