+++ inc/portf.h
@@ -13,11 +13,11 @@
 // Top level portoflio routines.
 bool FPortGetFniMovieOpen(FNI *pfni);
 bool FPortDisplayWithIds(FNI *pfni, bool fOpen, long lFilterLabel, long lFilterExt, long lTitle, LPCTSTR lpstrDefExt,
-                         PSTN pstnDefFileName, FNI *pfniInitialDir, ulong grfPrevType, CNO cnoWave);
-bool FPortGetFniOpen(FNI *pfni, LPCTSTR lpstrFilter, LPCTSTR lpstrTitle, FNI *pfniInitialDir, ulong grfPrevType,
+                         PSTN pstnDefFileName, FNI *pfniInitialDir, uint32_t grfPrevType, CNO cnoWave);
+bool FPortGetFniOpen(FNI *pfni, LPCTSTR lpstrFilter, LPCTSTR lpstrTitle, FNI *pfniInitialDir, uint32_t grfPrevType,
                      CNO cnoWave);
 bool FPortGetFniSave(FNI *pfni, LPCTSTR lpstrFilter, LPCTSTR lpstrTitle, LPCTSTR lpstrDefExt, PSTN pstnDefFileName,
-                     ulong grfPrevType, CNO cnoWave);
+                     uint32_t grfPrevType, CNO cnoWave);
 
 UINT_PTR CALLBACK OpenHookProc(HWND hWnd, UINT msg, WPARAM wParam, LPARAM lParam);
 void OpenPreview(HWND hwnd, PGNV pgnvOff, RCS *prcsPreview);
@@ -35,7 +35,7 @@ typedef struct dlginfo
     bool fIsOpen;      // fTrue if Open file, (ie not Save file)
     bool fDrawnBkgnd;  // fTrue if portfolio background bitmap has been displayed.
     RCS rcsDlg;        // Initial size of the portfolio common dlg window client area.
-    ulong grfPrevType; // Bits for types of preview required, (eg movie, sound etc) == 0 if no preview
+    uint32_t grfPrevType; // Bits for types of preview required, (eg movie, sound etc) == 0 if no preview
     CNO cnoWave;       // Wave file cno for audio when portfolio is invoked.
 } DLGINFO;
 typedef DLGINFO *PDLGINFO;
